% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{makepkg}
\alias{makepkg}
\title{Create PKGBUILD for a given R Package (Based on its DESCRIPTION file).}
\usage{
makepkg(file_description, file_output)
}
\arguments{
\item{file_description}{DESCRIPTION file path.}

\item{file_output}{PKGBUILD output file path.}
}
\value{
Path where the PKGBUILD was saved.
}
\description{
This function parses a DESCRIPTION file and creates a PKGBUILD.
}
\note{
To learn more about the PKGBUILD file format, please, check the
R package guidelines (https://wiki.archlinux.org/title/R_package_guidelines).
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
